+++ tests/unit/TelegramTest.php
@@ -112,36 +112,36 @@ class TelegramTest extends TestCase
     {
         $tg = $this->telegram;
 
-        $this->assertCount(1, $tg->getCommandsPaths());
+        $this->assertCount(0, $tg->getCommandsPaths());
 
         $tg->addCommandsPath($this->custom_commands_paths[0]);
-        $this->assertCount(2, $tg->getCommandsPaths());
+        $this->assertCount(1, $tg->getCommandsPaths());
         $this->assertArraySubset(
             [$this->custom_commands_paths[0]],
             $tg->getCommandsPaths()
         );
 
         $tg->addCommandsPath('/invalid/path');
-        $this->assertCount(2, $tg->getCommandsPaths());
+        $this->assertCount(1, $tg->getCommandsPaths());
 
         $tg->addCommandsPaths([
             $this->custom_commands_paths[1],
             $this->custom_commands_paths[2],
         ]);
-        $this->assertCount(4, $tg->getCommandsPaths());
+        $this->assertCount(3, $tg->getCommandsPaths());
         $this->assertArraySubset(
             array_reverse($this->custom_commands_paths),
             $tg->getCommandsPaths()
         );
 
         $tg->addCommandsPath($this->custom_commands_paths[0]);
-        $this->assertCount(4, $tg->getCommandsPaths());
+        $this->assertCount(3, $tg->getCommandsPaths());
     }
 
     public function testGetCommandsList()
     {
         $commands = $this->telegram->getCommandsList();
         $this->assertInternalType('array', $commands);
-        $this->assertNotCount(0, $commands);
+        $this->assertCount(0, $commands);
     }
 }